Moved files internally for better organization (#863)
