+++ src/io/parquet/read/deserialize/mod.rs
@@ -1,494 +1,271 @@
 //! APIs to read from Parquet format.
-use std::sync::Arc;
-
-use parquet2::{
-    schema::types::{
-        LogicalType, ParquetType, PhysicalType, TimeUnit as ParquetTimeUnit, TimestampType,
-    },
-    types::int96_to_i64_ns,
-};
+mod binary;
+mod boolean;
+mod dictionary;
+mod fixed_size_binary;
+mod nested_utils;
+mod null;
+mod primitive;
+mod simple;
+mod utils;
 
 use crate::{
-    array::{Array, BinaryArray, DictionaryKey, PrimitiveArray, Utf8Array},
-    datatypes::{DataType, IntervalUnit, TimeUnit},
+    array::{Array, BinaryArray, ListArray, StructArray, Utf8Array},
+    datatypes::{DataType, Field},
     error::{ArrowError, Result},
 };
 
-use super::binary;
-use super::boolean;
-use super::fixed_size_binary;
-use super::null;
-use super::primitive;
-use super::{dyn_iter, iden, op, ArrayIter, DataPages};
+use self::nested_utils::{InitNested, NestedArrayIter, NestedState};
+use simple::page_iter_to_arrays;
 
-pub fn page_iter_to_arrays<'a, I: 'a + DataPages>(
-    pages: I,
-    type_: &ParquetType,
+use super::*;
+
+/// Creates a new iterator of compressed pages.
+pub fn get_page_iterator<R: Read + Seek>(
+    column_metadata: &ColumnChunkMetaData,
+    reader: R,
+    pages_filter: Option<PageFilter>,
+    buffer: Vec<u8>,
+) -> Result<PageIterator<R>> {
+    Ok(_get_page_iterator(
+        column_metadata,
+        reader,
+        pages_filter,
+        buffer,
+    )?)
+}
+
+fn create_list(
     data_type: DataType,
-    chunk_size: usize,
-) -> Result<ArrayIter<'a>> {
-    use DataType::*;
+    nested: &mut NestedState,
+    values: Arc<dyn Array>,
+) -> Result<Arc<dyn Array>> {
+    Ok(match data_type {
+        DataType::List(_) => {
+            let (offsets, validity) = nested.nested.pop().unwrap().inner();
 
-    let (physical_type, logical_type) = if let ParquetType::PrimitiveType {
-        physical_type,
-        logical_type,
-        ..
-    } = type_
-    {
-        (physical_type, logical_type)
-    } else {
-        return Err(ArrowError::InvalidArgumentError(
-            "page_iter_to_arrays can only be called with a parquet primitive type".into(),
-        ));
-    };
-
-    Ok(match data_type.to_logical_type() {
-        Null => null::iter_to_arrays(pages, data_type, chunk_size),
-        Boolean => dyn_iter(boolean::Iter::new(pages, data_type, chunk_size)),
-        UInt8 => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: i32| x as u8,
-        ))),
-        UInt16 => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: i32| x as u16,
-        ))),
-        UInt32 => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: i32| x as u32,
-        ))),
-        Int8 => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: i32| x as i8,
-        ))),
-        Int16 => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: i32| x as i16,
-        ))),
-        Int32 | Date32 | Time32(_) | Interval(IntervalUnit::YearMonth) => dyn_iter(iden(
-            primitive::Iter::new(pages, data_type, chunk_size, |x: i32| x as i32),
-        )),
-
-        Timestamp(time_unit, _) => {
-            let time_unit = *time_unit;
-            return timestamp(
-                pages,
-                physical_type,
-                logical_type,
+            let offsets = offsets.iter().map(|x| *x as i32).collect::<Vec<_>>();
+            Arc::new(ListArray::<i32>::from_data(
                 data_type,
-                chunk_size,
-                time_unit,
-            );
+                offsets.into(),
+                values,
+                validity,
+            ))
         }
+        DataType::LargeList(_) => {
+            let (offsets, validity) = nested.nested.pop().unwrap().inner();
 
-        FixedSizeBinary(_) => dyn_iter(fixed_size_binary::Iter::new(pages, data_type, chunk_size)),
-
-        Decimal(_, _) => match physical_type {
-            PhysicalType::Int32 => dyn_iter(iden(primitive::Iter::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i32| x as i128,
-            ))),
-            PhysicalType::Int64 => dyn_iter(iden(primitive::Iter::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x as i128,
-            ))),
-            &PhysicalType::FixedLenByteArray(n) if n > 16 => {
-                return Err(ArrowError::NotYetImplemented(format!(
-                    "Can't decode Decimal128 type from Fixed Size Byte Array of len {:?}",
-                    n
-                )))
-            }
-            &PhysicalType::FixedLenByteArray(n) => {
-                let n = n as usize;
-
-                let pages =
-                    fixed_size_binary::Iter::new(pages, DataType::FixedSizeBinary(n), chunk_size);
-
-                let pages = pages.map(move |maybe_array| {
-                    let array = maybe_array?;
-                    let values = array
-                        .values()
-                        .chunks_exact(n)
-                        .map(|value: &[u8]| {
-                            // Copy the fixed-size byte value to the start of a 16 byte stack
-                            // allocated buffer, then use an arithmetic right shift to fill in
-                            // MSBs, which accounts for leading 1's in negative (two's complement)
-                            // values.
-                            let mut bytes = [0u8; 16];
-                            bytes[..n].copy_from_slice(value);
-                            i128::from_be_bytes(bytes) >> (8 * (16 - n))
-                        })
-                        .collect::<Vec<_>>();
-                    let validity = array.validity().cloned();
-
-                    Ok(PrimitiveArray::<i128>::from_data(
-                        data_type.clone(),
-                        values.into(),
-                        validity,
-                    ))
-                });
-
-                let arrays = pages.map(|x| x.map(|x| Arc::new(x) as Arc<dyn Array>));
-
-                Box::new(arrays) as _
-            }
-            _ => unreachable!(),
-        },
-
-        // INT64
-        Int64 | Date64 | Time64(_) | Duration(_) => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: i64| x as i64,
-        ))),
-        UInt64 => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: i64| x as u64,
-        ))),
-
-        Float32 => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: f32| x,
-        ))),
-        Float64 => dyn_iter(iden(primitive::Iter::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: f64| x,
-        ))),
-
-        Binary => dyn_iter(binary::Iter::<i32, BinaryArray<i32>, _>::new(
-            pages, data_type, chunk_size,
-        )),
-        LargeBinary => dyn_iter(binary::Iter::<i64, BinaryArray<i64>, _>::new(
-            pages, data_type, chunk_size,
-        )),
-        Utf8 => dyn_iter(binary::Iter::<i32, Utf8Array<i32>, _>::new(
-            pages, data_type, chunk_size,
-        )),
-        LargeUtf8 => dyn_iter(binary::Iter::<i64, Utf8Array<i64>, _>::new(
-            pages, data_type, chunk_size,
-        )),
-
-        Dictionary(key_type, _, _) => {
-            return match_integer_type!(key_type, |$K| {
-                dict_read::<$K, _>(pages, physical_type, logical_type, data_type, chunk_size)
-            })
+            Arc::new(ListArray::<i64>::from_data(
+                data_type, offsets, values, validity,
+            ))
         }
-
-        other => {
+        _ => {
             return Err(ArrowError::NotYetImplemented(format!(
-                "Reading {:?} from parquet still not implemented",
-                other
+                "Read nested datatype {:?}",
+                data_type
             )))
         }
     })
 }
 
-fn timestamp<'a, I: 'a + DataPages>(
-    pages: I,
-    physical_type: &PhysicalType,
-    logical_type: &Option<LogicalType>,
-    data_type: DataType,
-    chunk_size: usize,
-    time_unit: TimeUnit,
-) -> Result<ArrayIter<'a>> {
-    if physical_type == &PhysicalType::Int96 {
-        if time_unit == TimeUnit::Nanosecond {
-            return Ok(dyn_iter(iden(primitive::Iter::new(
-                pages,
-                data_type,
-                chunk_size,
-                int96_to_i64_ns,
-            ))));
-        } else {
-            return Err(ArrowError::nyi(
-                "Can't decode int96 to timestamp other than ns",
-            ));
+struct StructIterator<'a> {
+    iters: Vec<NestedArrayIter<'a>>,
+    fields: Vec<Field>,
+}
+
+impl<'a> StructIterator<'a> {
+    pub fn new(iters: Vec<NestedArrayIter<'a>>, fields: Vec<Field>) -> Self {
+        assert_eq!(iters.len(), fields.len());
+        Self { iters, fields }
+    }
+}
+
+impl<'a> Iterator for StructIterator<'a> {
+    type Item = Result<(NestedState, Arc<dyn Array>)>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let values = self
+            .iters
+            .iter_mut()
+            .map(|iter| iter.next())
+            .collect::<Vec<_>>();
+
+        if values.iter().any(|x| x.is_none()) {
+            return None;
         }
-    };
-    if physical_type != &PhysicalType::Int64 {
-        return Err(ArrowError::nyi(
-            "Can't decode a timestamp from a non-int64 parquet type",
+        let values = values
+            .into_iter()
+            .map(|x| x.unwrap().map(|x| x.1))
+            .collect::<Result<Vec<_>>>();
+
+        match values {
+            Ok(values) => Some(Ok((
+                NestedState::new(vec![]), // todo
+                Arc::new(StructArray::from_data(
+                    DataType::Struct(self.fields.clone()),
+                    values,
+                    None,
+                )),
+            ))),
+            Err(e) => Some(Err(e)),
+        }
+    }
+}
+
+fn columns_to_iter_recursive<'a, I: 'a>(
+    mut columns: Vec<I>,
+    mut types: Vec<&ParquetType>,
+    field: Field,
+    mut init: Vec<InitNested>,
+    chunk_size: usize,
+) -> Result<NestedArrayIter<'a>>
+where
+    I: DataPages,
+{
+    use DataType::*;
+    if init.len() == 1 && init[0].is_primitive() {
+        return Ok(Box::new(
+            page_iter_to_arrays(
+                columns.pop().unwrap(),
+                types.pop().unwrap(),
+                field.data_type,
+                chunk_size,
+            )?
+            .map(|x| Ok((NestedState::new(vec![]), x?))),
         ));
     }
 
-    let iter = primitive::Iter::new(pages, data_type, chunk_size, |x: i64| x);
-
-    let unit = if let Some(LogicalType::TIMESTAMP(TimestampType { unit, .. })) = logical_type {
-        unit
-    } else {
-        return Ok(dyn_iter(iden(iter)));
-    };
-
-    Ok(match (unit, time_unit) {
-        (ParquetTimeUnit::MILLIS(_), TimeUnit::Second) => dyn_iter(op(iter, |x| x / 1_000)),
-        (ParquetTimeUnit::MICROS(_), TimeUnit::Second) => dyn_iter(op(iter, |x| x / 1_000_000)),
-        (ParquetTimeUnit::NANOS(_), TimeUnit::Second) => dyn_iter(op(iter, |x| x * 1_000_000_000)),
-
-        (ParquetTimeUnit::MILLIS(_), TimeUnit::Millisecond) => dyn_iter(iden(iter)),
-        (ParquetTimeUnit::MICROS(_), TimeUnit::Millisecond) => dyn_iter(op(iter, |x| x / 1_000)),
-        (ParquetTimeUnit::NANOS(_), TimeUnit::Millisecond) => dyn_iter(op(iter, |x| x / 1_000_000)),
-
-        (ParquetTimeUnit::MILLIS(_), TimeUnit::Microsecond) => dyn_iter(op(iter, |x| x * 1_000)),
-        (ParquetTimeUnit::MICROS(_), TimeUnit::Microsecond) => dyn_iter(iden(iter)),
-        (ParquetTimeUnit::NANOS(_), TimeUnit::Microsecond) => dyn_iter(op(iter, |x| x / 1_000)),
-
-        (ParquetTimeUnit::MILLIS(_), TimeUnit::Nanosecond) => dyn_iter(op(iter, |x| x * 1_000_000)),
-        (ParquetTimeUnit::MICROS(_), TimeUnit::Nanosecond) => dyn_iter(op(iter, |x| x * 1_000)),
-        (ParquetTimeUnit::NANOS(_), TimeUnit::Nanosecond) => dyn_iter(iden(iter)),
+    Ok(match field.data_type().to_logical_type() {
+        Boolean => {
+            types.pop();
+            boolean::iter_to_arrays_nested(columns.pop().unwrap(), init.pop().unwrap(), chunk_size)
+        }
+        Int16 => {
+            types.pop();
+            primitive::iter_to_arrays_nested(
+                columns.pop().unwrap(),
+                init.pop().unwrap(),
+                field.data_type().clone(),
+                chunk_size,
+                |x: i32| x as i16,
+            )
+        }
+        Int64 => {
+            types.pop();
+            primitive::iter_to_arrays_nested(
+                columns.pop().unwrap(),
+                init.pop().unwrap(),
+                field.data_type().clone(),
+                chunk_size,
+                |x: i64| x,
+            )
+        }
+        Utf8 => {
+            types.pop();
+            binary::iter_to_arrays_nested::<i32, Utf8Array<i32>, _>(
+                columns.pop().unwrap(),
+                init.pop().unwrap(),
+                field.data_type().clone(),
+                chunk_size,
+            )
+        }
+        LargeBinary => {
+            types.pop();
+            binary::iter_to_arrays_nested::<i64, BinaryArray<i64>, _>(
+                columns.pop().unwrap(),
+                init.pop().unwrap(),
+                field.data_type().clone(),
+                chunk_size,
+            )
+        }
+        List(inner) => {
+            let iter = columns_to_iter_recursive(
+                vec![columns.pop().unwrap()],
+                types,
+                inner.as_ref().clone(),
+                init,
+                chunk_size,
+            )?;
+            let iter = iter.map(move |x| {
+                let (mut nested, array) = x?;
+                let array = create_list(field.data_type().clone(), &mut nested, array)?;
+                Ok((nested, array))
+            });
+            Box::new(iter) as _
+        }
+        Struct(fields) => {
+            let columns = fields
+                .iter()
+                .rev()
+                .map(|f| {
+                    columns_to_iter_recursive(
+                        vec![columns.pop().unwrap()],
+                        vec![types.pop().unwrap()],
+                        f.clone(),
+                        vec![init.pop().unwrap()],
+                        chunk_size,
+                    )
+                })
+                .collect::<Result<Vec<_>>>()?;
+            let columns = columns.into_iter().rev().collect();
+            Box::new(StructIterator::new(columns, fields.clone()))
+        }
+        _ => todo!(),
     })
 }
 
-fn timestamp_dict<'a, K: DictionaryKey, I: 'a + DataPages>(
-    pages: I,
-    physical_type: &PhysicalType,
-    logical_type: &Option<LogicalType>,
-    data_type: DataType,
-    chunk_size: usize,
-    time_unit: TimeUnit,
-) -> Result<ArrayIter<'a>> {
-    if physical_type == &PhysicalType::Int96 {
-        if time_unit == TimeUnit::Nanosecond {
-            return Ok(dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                DataType::Timestamp(TimeUnit::Nanosecond, None),
-                chunk_size,
-                int96_to_i64_ns,
-            )));
-        } else {
-            return Err(ArrowError::nyi(
-                "Can't decode int96 to timestamp other than ns",
-            ));
+fn field_to_init(field: &Field) -> Vec<InitNested> {
+    use crate::datatypes::PhysicalType::*;
+    match field.data_type.to_physical_type() {
+        Null | Boolean | Primitive(_) | Binary | FixedSizeBinary | LargeBinary | Utf8
+        | Dictionary(_) | LargeUtf8 => vec![InitNested::Primitive(field.is_nullable)],
+        List | FixedSizeList | LargeList => {
+            let a = field.data_type().to_logical_type();
+            let inner = if let DataType::List(inner) = a {
+                field_to_init(inner)
+            } else if let DataType::LargeList(inner) = a {
+                field_to_init(inner)
+            } else if let DataType::FixedSizeList(inner, _) = a {
+                field_to_init(inner)
+            } else {
+                unreachable!()
+            };
+            inner
+                .into_iter()
+                .map(|x| InitNested::List(Box::new(x), field.is_nullable))
+                .collect()
         }
-    };
-
-    let unit = if let Some(LogicalType::TIMESTAMP(TimestampType { unit, .. })) = logical_type {
-        unit
-    } else {
-        return Ok(dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-            pages,
-            data_type,
-            chunk_size,
-            |x: i64| x,
-        )));
-    };
-
-    Ok(match (unit, time_unit) {
-        (ParquetTimeUnit::MILLIS(_), TimeUnit::Second) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x / 1_000,
-            ))
+        Struct => {
+            let inner = if let DataType::Struct(fields) = field.data_type.to_logical_type() {
+                fields.iter().rev().map(field_to_init).collect::<Vec<_>>()
+            } else {
+                unreachable!()
+            };
+            inner
+                .into_iter()
+                .flatten()
+                .map(|x| InitNested::Struct(Box::new(x), field.is_nullable))
+                .collect()
         }
-        (ParquetTimeUnit::MICROS(_), TimeUnit::Second) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x / 1_000_000,
-            ))
-        }
-        (ParquetTimeUnit::NANOS(_), TimeUnit::Second) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x * 1_000_000_000,
-            ))
-        }
-
-        (ParquetTimeUnit::MILLIS(_), TimeUnit::Millisecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x,
-            ))
-        }
-        (ParquetTimeUnit::MICROS(_), TimeUnit::Millisecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x / 1_000,
-            ))
-        }
-        (ParquetTimeUnit::NANOS(_), TimeUnit::Millisecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x / 1_000_000,
-            ))
-        }
-
-        (ParquetTimeUnit::MILLIS(_), TimeUnit::Microsecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x * 1_000,
-            ))
-        }
-        (ParquetTimeUnit::MICROS(_), TimeUnit::Microsecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x,
-            ))
-        }
-        (ParquetTimeUnit::NANOS(_), TimeUnit::Microsecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x / 1_000,
-            ))
-        }
-
-        (ParquetTimeUnit::MILLIS(_), TimeUnit::Nanosecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x * 1_000_000,
-            ))
-        }
-        (ParquetTimeUnit::MICROS(_), TimeUnit::Nanosecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x * 1_000,
-            ))
-        }
-        (ParquetTimeUnit::NANOS(_), TimeUnit::Nanosecond) => {
-            dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-                pages,
-                data_type,
-                chunk_size,
-                |x: i64| x,
-            ))
-        }
-    })
+        _ => todo!(),
+    }
 }
 
-fn dict_read<'a, K: DictionaryKey, I: 'a + DataPages>(
-    iter: I,
-    physical_type: &PhysicalType,
-    logical_type: &Option<LogicalType>,
-    data_type: DataType,
+/// An iterator adapter that maps multiple iterators of [`DataPages`] into an iterator of [`Array`]s.
+///
+/// The arrays are guaranteed to be at most of size `chunk_size` and data type `field.data_type`.
+pub fn column_iter_to_arrays<'a, I: 'a>(
+    columns: Vec<I>,
+    types: Vec<&ParquetType>,
+    field: Field,
     chunk_size: usize,
-) -> Result<ArrayIter<'a>> {
-    use DataType::*;
-    let values_data_type = if let Dictionary(_, v, _) = &data_type {
-        v.as_ref()
-    } else {
-        panic!()
-    };
+) -> Result<ArrayIter<'a>>
+where
+    I: DataPages,
+{
+    let init = field_to_init(&field);
 
-    Ok(match values_data_type.to_logical_type() {
-        UInt8 => dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-            iter,
-            data_type,
-            chunk_size,
-            |x: i32| x as u8,
-        )),
-        UInt16 => dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-            iter,
-            data_type,
-            chunk_size,
-            |x: i32| x as u16,
-        )),
-        UInt32 => dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-            iter,
-            data_type,
-            chunk_size,
-            |x: i32| x as u32,
-        )),
-        Int8 => dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-            iter,
-            data_type,
-            chunk_size,
-            |x: i32| x as i8,
-        )),
-        Int16 => dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-            iter,
-            data_type,
-            chunk_size,
-            |x: i32| x as i16,
-        )),
-        Int32 | Date32 | Time32(_) | Interval(IntervalUnit::YearMonth) => dyn_iter(
-            primitive::DictIter::<K, _, _, _, _>::new(iter, data_type, chunk_size, |x: i32| {
-                x as i32
-            }),
-        ),
-
-        Timestamp(time_unit, _) => {
-            let time_unit = *time_unit;
-            return timestamp_dict::<K, _>(
-                iter,
-                physical_type,
-                logical_type,
-                data_type,
-                chunk_size,
-                time_unit,
-            );
-        }
-
-        Int64 | Date64 | Time64(_) | Duration(_) => dyn_iter(
-            primitive::DictIter::<K, _, _, _, _>::new(iter, data_type, chunk_size, |x: i64| x),
-        ),
-        Float32 => dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-            iter,
-            data_type,
-            chunk_size,
-            |x: f32| x,
-        )),
-        Float64 => dyn_iter(primitive::DictIter::<K, _, _, _, _>::new(
-            iter,
-            data_type,
-            chunk_size,
-            |x: f64| x,
-        )),
-
-        Utf8 | Binary => dyn_iter(binary::DictIter::<K, i32, _>::new(
-            iter, data_type, chunk_size,
-        )),
-        LargeUtf8 | LargeBinary => dyn_iter(binary::DictIter::<K, i64, _>::new(
-            iter, data_type, chunk_size,
-        )),
-        FixedSizeBinary(_) => dyn_iter(fixed_size_binary::DictIter::<K, _>::new(
-            iter, data_type, chunk_size,
-        )),
-        other => {
-            return Err(ArrowError::nyi(format!(
-                "Reading dictionaries of type {:?}",
-                other
-            )))
-        }
-    })
+    Ok(Box::new(
+        columns_to_iter_recursive(columns, types, field, init, chunk_size)?.map(|x| x.map(|x| x.1)),
+    ))
 }
